Add PAuth helpers for signing, stripping and authentication of pointers
Add UC_CTL_PAUTH_SIGN, UC_CTL_PAUTH_STRIP and UC_CTL_PAUTH_AUTH as an
architecture-independent interface to pointer authentication operations
without having to execute instructions on the virtual CPU.  This is
useful in many scenarios, for instance when rebasing signed pointers as
part of loading code that runs with PAuth.

The C interfaces are covered by a unit test.  The Python interfaces have
been tested in a real world project.